Airflow_Lab/requirements.txt
pandas
scikit-learn
matplotlib
apache-airflow
›